Move producer/data producer closing integration tests for consumers/data consumers to unit tests so that we can actually close producers using test-only library API
