+++ rust/tests/integration/consumer.rs
@@ -1402,42 +1402,3 @@ fn close_event() {
         }
     });
 }
-
-#[test]
-fn producer_close_event() {
-    future::block_on(async move {
-        let (_worker, _router, transport_1, transport_2) = init().await;
-
-        let audio_producer = transport_1
-            .produce(audio_producer_options())
-            .await
-            .expect("Failed to produce audio");
-
-        let audio_consumer = transport_2
-            .consume(ConsumerOptions::new(
-                audio_producer.id(),
-                consumer_device_capabilities(),
-            ))
-            .await
-            .expect("Failed to consume audio");
-
-        let (mut close_tx, close_rx) = async_oneshot::oneshot::<()>();
-        let _handler = audio_consumer.on_close(move || {
-            let _ = close_tx.send(());
-        });
-
-        let (mut producer_close_tx, producer_close_rx) = async_oneshot::oneshot::<()>();
-        let _handler = audio_consumer.on_producer_close(move || {
-            let _ = producer_close_tx.send(());
-        });
-        drop(audio_producer);
-
-        producer_close_rx
-            .await
-            .expect("Failed to receive producer_close event");
-
-        close_rx.await.expect("Failed to receive close event");
-
-        assert_eq!(audio_consumer.closed(), true);
-    });
-}
